break ListingsFullDetailsPage component codes into sub-components ContentSidebar ContentMain
